requirements.txt
click>=8.0.0,<9.0.0
google-analytics-admin>=0.27.0,<1.0.0
google-auth>=2.0.0,<3.0.0
rich>=13.0.0,<14.0.0
pytz>=2023.3
